Don't try to delete file when it doesn't exist

+++ identity/identityeditvcarddialog.cpp
@@ -61,8 +61,10 @@ void IdentityEditVcardDialog::deleteCurrentVcard()
 {
     if (!mVcardFileName.isEmpty()) {
         QFile file(mVcardFileName);
-        if (!file.remove()) {
-            KMessageBox::error(this, i18n("We can not delete vcard file."), i18n("Delete vcard"));
+        if (file.exists()) {
+            if (!file.remove()) {
+                KMessageBox::error(this, i18n("We can not delete vcard file."), i18n("Delete vcard"));
+            }
         }
     }
 }
